+++ src/nrf_802154_ack_pending_bit.c
@@ -170,12 +170,12 @@ static bool addr_binary_search(const uint8_t * p_addr,
     uint8_t addr_array_len = extended ? m_num_of_pending_extended : m_num_of_pending_short;
     uint8_t entry_size     = extended ? EXTENDED_ADDRESS_SIZE : SHORT_ADDRESS_SIZE;
     int8_t  low            = 0;
-    int8_t  midpoint       = 0;
+    uint8_t midpoint       = 0;
     int8_t  high           = addr_array_len;
 
     while (high >= low)
     {
-        midpoint = low + (high - low) / 2;
+        midpoint = (uint8_t)(low + (high - low) / 2);
 
         if (midpoint >= addr_array_len)
         {
@@ -185,7 +185,7 @@ static bool addr_binary_search(const uint8_t * p_addr,
         switch (addr_compare(p_addr, p_addr_array + entry_size * midpoint, extended))
         {
             case -1:
-                high = midpoint - 1;
+                high = (int8_t)(midpoint - 1);
                 break;
 
             case 0:
@@ -193,7 +193,7 @@ static bool addr_binary_search(const uint8_t * p_addr,
                 return true;
 
             case 1:
-                low = midpoint + 1;
+                low = (int8_t)(midpoint + 1);
                 break;
 
             default:
@@ -209,7 +209,7 @@ static bool addr_binary_search(const uint8_t * p_addr,
      * to 0 and with last case being utilized, low is set to 1. However, midpoint equal to 0 is
      * incorrect, as in the last iteration first element of the array proves to be less than the
      * element to be added to the array. With the below code, midpoint is then shifted to 1. */
-    if (low == midpoint + 1)
+    if ((uint8_t)low == midpoint + 1)
     {
         midpoint++;
     }
